Module sprs::binop

Sparse matrix addition, subtraction

Functions

add_dense_mat_same_ordering
Compute alpha * lhs + beta * rhs with lhs a sparse matrix and rhs dense and alpha and beta scalars

csmat_binop
Applies a binary operation to matching non-zero elements of two sparse matrices. When e.g. only the lhs has a non-zero at a given location, 0 is inferred for the non-zero value of the other matrix. Both matrices should have the same storage.

csmat_binop_dense_raw
Raw implementation of sparse/dense binary operations with the same ordering

csmat_binop_same_storage_raw
Raw implementation of scalar binary operation for compressed sparse matrices sharing the same storage. The output arrays are assumed to be preallocated

csvec_binop
Binary operations for CsVec

mul_dense_mat_same_ordering
Compute coeff wise alpha * lhs * rhs with lhs a sparse matrix, rhs a dense matrix, and alpha a scalar

mul_mat_same_storage
Sparse matrix scalar multiplication, with same storage type