Function sprs::linalg::trisolve::lsolve_csr_dense_rhs

pub fn lsolve_csr_dense_rhs<N, I, Iptr, V>(
    lower_tri_mat: CsMatViewI<'_, N, I, Iptr>, 
    rhs: V
) -> Result<(), LinalgError> where
    N: Clone + Num + SubAssign,
    &'r N: Mul<&'r N, Output = N>,
    V: DenseVectorMut<Scalar = N>,
    I: SpIndex,
    Iptr: SpIndex, 

Solve a sparse lower triangular matrix system, with a csr matrix and a dense vector as inputs

The solve results are written into the provided values.

This solve does not assume the input matrix to actually be triangular, instead it ignores the upper triangular part.