Module sprs::sparse::binop

Functions

add_dense_mat_same_ordering
Compute alpha * lhs + beta * rhs with lhs a sparse matrix and rhs dense and alpha and beta scalars

add_mat_same_storage
Sparse matrix addition, with matrices sharing the same storage type

csmat_binop_dense_same_ordering_raw
Raw implementation of sparse/dense binary operations with the same ordering

csmat_binop_same_storage_raw
Raw implementation of scalar binary operation for compressed sparse matrices sharing the same storage. The output arrays are assumed to be preallocated

csvec_binop
Binary operations for CsVec

mul_dense_mat_same_ordering
Compute coeff wise alpha * lhs * rhs with lhs a sparse matrix and rhs dense and alpha a scalar

mul_mat_same_storage
Sparse matrix scalar multiplication, with same storage type

scalar_mul_mat
Sparse matrix multiplication by a scalar

scalar_mul_mat_raw
Sparse matrix multiplication by a scalar, raw version

sub_mat_same_storage
Sparse matrix subtraction, with same storage type