Module sprs::sparse::construct

High level construction of sparse matrices by stacking, by block, ...

Functions

bmat
Specify a sparse matrix by constructing it from blocks of other matrices

hstack
Construct a sparse matrix by horizontally stacking other matrices

same_storage_fast_stack
Stack the given matrices into a new one, using the most efficient stacking direction (ie vertical stack for CSR matrices, horizontal stack for CSC)

vstack
Construct a sparse matrix by vertically stacking other matrices