Module sprs::sparse::csmat

Structs

ChunkOuterBlocks
An iterator over non-overlapping blocks of a matrix, along the least-varying dimension

CsMat
Compressed matrix in the CSR or CSC format.

OuterIterator
Iterator on the matrix' outer dimension Implemented over an iterator on the indptr array

OuterIteratorPerm
Iterator on the matrix' outer dimension, permuted Implemented over an iterator on the indptr array

Enums

CompressedStorage
Describe the storage of a CsMat

Type Definitions

CsMatOwned

CsMatVecView

CsMatView