Struct sprs::sparse::linalg::etree::Parents

pub struct Parents<S> where S: Deref<Target=[Parent]> {
    // some fields omitted
}

Store an etree as the parent information of each node. This reflects the fact that etrees can in fact have multiple roots.

Methods

impl<S: Deref<Target=[Parent]>> Parents<S>

fn get_parent(&self, node: usize) -> Option<usize>

Get the parent of a node. Returns None if the node is a root.

Panics

if node is out of bounds

fn is_root(&self, node: usize) -> bool

Test whether a node is a root.

Panics

if node is out of bounds

fn nb_nodes(&self) -> usize

The number of nodes in this tree.

fn borrowed(&self) -> ParentsView

impl<S: DerefMut<Target=[Parent]>> Parents<S>

fn set_parent(&mut self, node: usize, parent: usize)

Set the parent of a node.

Panics

if node is out of bounds

if parent is out of bounds

fn set_root(&mut self, node: usize)

Set a node as a root.

Panics

if node is out of bounds

fn uproot(&mut self, node: usize, parent: usize)

Give a parent to a root of the tree. No-op if the node was not a parent.

Panics

if either node or parent is out of bounds

fn borrowed_mut(&mut self) -> ParentsViewMut

impl Parents<Vec<Parent>>

fn new(nb_nodes: usize) -> ParentsOwned

Create a new tree with all nodes set as root