Module sprs::sparse::prod

Functions

csc_mul_csc
Perform a matrix multiplication for matrices sharing the same storage order.

csc_mulacc_dense_colmaj
CSC-dense colmaj multiplication

csc_mulacc_dense_rowmaj
CSC-dense rowmaj multiplication

csr_mul_csr
Perform a matrix multiplication for matrices sharing the same storage order.

csr_mul_csr_impl
Actual implementation of CSR-CSR multiplication All other matrix products are implemented in terms of this one.

csr_mul_csvec
CSR-vector multiplication

csr_mulacc_dense_colmaj
CSR-dense colmaj multiplication

csr_mulacc_dense_rowmaj
CSR-dense rowmaj multiplication

mul_acc_mat_vec_csc
Multiply a sparse CSC matrix with a dense vector and accumulate the result into another dense vector

mul_acc_mat_vec_csr
Multiply a sparse CSR matrix with a dense vector and accumulate the result into another dense vector

workspace_csc
Allocate the appropriate workspace for a CSC-CSC product

workspace_csr
Allocate the appropriate workspace for a CSR-CSR product