Function sprs::sparse::prod::csc_mul_csc

pub fn csc_mul_csc<N, Mat1, Mat2>(lhs: &Mat1, rhs: &Mat2, workspace: &mut [N]) -> Result<CsMatOwned<N>, SprsError> where N: Num + Copy, Mat1: SpMatView<N>, Mat2: SpMatView<N>

Perform a matrix multiplication for matrices sharing the same storage order.

This method assumes a CSC storage order, and uses free transposition to invoke the CSR method

lhs: left hand size matrix rhs: right hand size matrix workspace: used to accumulate the line values. Should be of length lhs.lines()