Struct sprs::sparse::triplet::TripletMat

pub struct TripletMat<N> {
    // some fields omitted
}

Triplet matrix owning its data

Methods

impl<N> TripletMat<N>

fn new(shape: (usize, usize)) -> TripletMat<N>

Create a new triplet matrix of shape (nb_rows, nb_cols)

fn with_capacity(shape: (usize, usize), cap: usize) -> TripletMat<N>

Create a new triplet matrix of shape (nb_rows, nb_cols), and pre-allocate cap elements on the backing storage

fn from_triplets(shape: (usize, usize), row_inds: Vec<usize>, col_inds: Vec<usize>, data: Vec<N>) -> TripletMat<N>

Create a triplet matrix from its raw components. All arrays should have the same length.

Panics

if the arrays don't have the same length

if either the row or column indices are out of bounds.

fn rows(&self) -> usize

The number of rows of the matrix

fn cols(&self) -> usize

The number of cols of the matrix

fn shape(&self) -> (usize, usize)

The shape of the matrix, as a (rows, cols) tuple

fn nnz(&self) -> usize

The number of non-zero entries

fn row_inds(&self) -> &[usize]

The non-zero row indices

fn col_inds(&self) -> &[usize]

The non-zero column indices

fn data(&self) -> &[N]

The non-zero values

fn find_locations(&self, row: usize, col: usize) -> Vec<TripletIndex>

Find all non-zero entries at the location given by row and col

fn borrowed(&self) -> TripletView<N>

Return a view of this matrix

fn set_triplet(&mut self, TripletIndex: TripletIndex, row: usize, col: usize, val: N)

Replace a non-zero value at the given index. Indices can be obtained using find_locations.

fn borrowed_mut(&mut self) -> TripletViewMut<N>

Get a mutable view into this matrix.

fn transpose_view(&self) -> TripletView<N>

Get a transposed view of this matrix

fn add_triplet(&mut self, row: usize, col: usize, val: N)

Append a non-zero triplet to this matrix.

fn reserve(&mut self, cap: usize)

Reserve cap additional non-zeros

fn reserve_exact(&mut self, cap: usize)

Reserve exactly cap non-zeros

fn to_csc(&self) -> CsMatOwned<N> where N: Copy + Num

Create a CSC matrix from this triplet matrix

fn to_csr(&self) -> CsMatOwned<N> where N: Copy + Num

Create a CSR matrix from this triplet matrix