Struct sprs::sparse::vec::CsVec

pub struct CsVec<N, IStorage, DStorage> where DStorage: Deref<Target=[N]> {
    // some fields omitted
}

A sparse vector, storing the indices of its non-zero data. The indices should be sorted.

Methods

impl<N> CsVec<N, Vec<usize>, Vec<N>>

fn new_owned(n: usize, indices: Vec<usize>, data: Vec<N>) -> Result<CsVec<N, Vec<usize>, Vec<N>>, SprsError>

Create an owning CsVec from vector data.

fn empty(dim: usize) -> CsVec<N, Vec<usize>, Vec<N>>

Create an empty CsVec, which can be used for incremental construction

fn append(&mut self, ind: usize, val: N)

Append an element to the sparse vector. Used for incremental building of the CsVec. The append should preserve the structure of the vector, ie the newly added index should be strictly greater than the last element of indices.

Panics

Panics if ind is lower or equal to the last element of self.indices() Panics if ind is greater than self.dim()

fn reserve(&mut self, size: usize)

Reserve size additional non-zero values.

fn reserve_exact(&mut self, exact_size: usize)

Reserve exactly exact_size non-zero values.

fn clear(&mut self)

Clear the underlying storage

impl<N, IStorage, DStorage> CsVec<N, IStorage, DStorage> where IStorage: Deref<Target=[usize]>, DStorage: Deref<Target=[N]>

fn borrowed(&self) -> CsVecView<N>

Get a view of this vector.

impl<N, IStorage, DStorage> CsVec<N, IStorage, DStorage> where IStorage: Deref<Target=[usize]>, DStorage: Deref<Target=[N]>

fn iter(&self) -> VectorIterator<N>

Iterate over the non zero values.

Example

use sprs::CsVec;
let v = CsVec::new_owned(5, vec![0, 2, 4], vec![1., 2., 3.]).unwrap();
let mut iter = v.iter();
assert_eq!(iter.next(), Some((0, &1.)));
assert_eq!(iter.next(), Some((2, &2.)));
assert_eq!(iter.next(), Some((4, &3.)));
assert_eq!(iter.next(), None);

fn indices(&self) -> &[usize]

The underlying indices.

fn data(&self) -> &[N]

The underlying non zero values.

fn dim(&self) -> usize

The dimension of this vector.

fn nnz(&self) -> usize

The non zero count of this vector.

fn check_structure(&self) -> Result<(), SprsError>

Check the sparse structure, namely that: - indices is sorted - indices are lower than dims()

fn to_owned(&self) -> CsVecOwned<N> where N: Clone

Allocate a new vector equal to this one.

fn row_view(&self) -> CsMatVecView<N>

View this vector as a matrix with only one row.

fn col_view(&self) -> CsMatVecView<N>

View this vector as a matrix with only one column.

fn get(&self, index: usize) -> Option<&N>

Access element at given index, with logarithmic complexity

fn nnz_index(&self, index: usize) -> Option<NnzIndex>

Find the non-zero index of the requested dimension index, returning None if no non-zero is present at the requested location.

Looking for the NnzIndex is done with logarithmic complexity, but once it is available, the NnzIndex enables retrieving the data with O(1) complexity.

fn dot<'b, T: IntoSparseVecIter<&'b N>>(&'b self, rhs: T) -> N where N: Num + Copy, T::IterType: Iterator<Item=(usize, &'b N)>

Sparse vector dot product. The right-hand-side can be any type that can be interpreted as a sparse vector (hence sparse vectors, std vectors and slices, and ndarray's dense vectors work).

Panics

If the dimension of the vectors do not match.

Example

use sprs::CsVec;
let v1 = CsVec::new_owned(8, vec![1, 2, 4, 6], vec![1.; 4]).unwrap();
let v2 = CsVec::new_owned(8, vec![1, 3, 5, 7], vec![2.; 4]).unwrap();
assert_eq!(2., v1.dot(&v2));
assert_eq!(4., v1.dot(&v1));
assert_eq!(16., v2.dot(&v2));

fn scatter(&self, out: &mut [N]) where N: Clone

Fill a dense vector with our values

fn to_set(self) -> HashSet<(usize, N)> where N: Hash + Eq + Clone

Transform this vector into a set of (index, value) tuples

fn map<F>(&self, f: F) -> CsVecOwned<N> where F: FnMut(&N) -> N, N: Clone

Apply a function to each non-zero element, yielding a new matrix with the same sparsity structure.

impl<'a, N, IStorage, DStorage> CsVec<N, IStorage, DStorage> where N: 'a, IStorage: 'a + Deref<Target=[usize]>, DStorage: DerefMut<Target=[N]>

fn borrowed_mut(&mut self) -> CsVecViewMut<N>

fn get_mut(&mut self, index: usize) -> Option<&mut N>

Access element at given index, with logarithmic complexity

fn map_inplace<F>(&mut self, f: F) where F: FnMut(&N) -> N

Apply a function to each non-zero element, mutating it