Trait sprs::sparse::vec::IntoSparseVecIter

pub trait IntoSparseVecIter<N> {
    type IterType;
    fn into_sparse_vec_iter(self) -> Self::IterType where Self::IterType: Iterator<Item=(usize, N)>;
    fn dim(&self) -> usize;
}

Trait for types that can be iterated as sparse vectors

Associated Types

type IterType

Required Methods

fn into_sparse_vec_iter(self) -> Self::IterType where Self::IterType: Iterator<Item=(usize, N)>

Transform self into an iterator that yields (usize, &N) tuples where the usize is the index of the value in the sparse vector. The indices should be sorted.

fn dim(&self) -> usize

The dimension of the vector