Struct sprs::stack::DStack

pub struct DStack<I> {
    // some fields omitted
}

A double stack of fixed capacity, holding recursion information (eg for dfs) as well as data values.

Used in sparse triangular / sparse vector solves, where it is guaranteed that the two parts of the stack cannot overlap.

Methods

impl<I> DStack<I> where I: Copy

fn with_capacity(n: usize) -> DStack<I> where I: Default

Create a new double stacked suited for containing at most n elements

fn capacity(&self) -> usize

Capacity of the dstack

fn is_rec_empty(&self) -> bool

Test whether the recursion stack is empty

fn is_data_empty(&self) -> bool

Test whether the data stack is empty

fn push_rec(&mut self, value: StackVal<I>)

Push a value on the recursion stack

fn push_data(&mut self, value: I)

Push a value on the data stack

fn pop_rec(&mut self) -> Option<StackVal<I>>

Pop a value from the recursion stack

fn pop_data(&mut self) -> Option<I>

Pop a value from the data stack

fn len_data(&self) -> usize

Number of data elements this double stack contains

fn clear_data(&mut self)

Clear the data stack

fn iter_data<'a>(&'a self) -> Map<Iter<'a, StackVal<I>>, fn(&StackVal<I>) -> &I>

Iterates along the data stack without removing items