Struct sprs::TriMatBase

pub struct TriMatBase<IStorage, DStorage> { /* fields omitted */ }

Sparse matrix in the triplet format.

Sparse matrices in the triplet format use three arrays of equal sizes (accessible through the methods row_inds, col_inds, data), the first one storing the row indices of non-zero values, the second storing the corresponding column indices and the last array storing the corresponding scalar value. If a non-zero location is repeated in the arrays, the non-zero value is taken as the sum of the corresponding scalar entries.

This format is useful for iteratively building a sparse matrix, since the various non-zero entries can be specified in any order, or even partially as is common in physics with partial derivatives equations.

This format cannot be used for arithmetic operations. Arithmetic operations are more efficient in the compressed format. A matrix in the triplet format can be converted to the compressed format using the methods to_csc and to_csr.

The TriMatBase type is parameterized by the storage type for the row and column indices, IStorage, and by the storage type for the non-zero values DStorage. Convenient aliases are availaible to specify frequent variant: TriMat refers to a triplet matrix owning the storage of its indices and and values, TriMatView refers to a triplet matrix with slices to store its indices and values, while TriMatViewMut refers to a a triplet matrix using mutable slices.

Additionaly, the type aliases TriMatI, TriMatViewI and TriMatViewMutI can be used to choose an index type different from the default usize.

Implementations

impl<N, I: SpIndex> TriMatBase<Vec<I, Global>, Vec<N, Global>>

Methods for creating triplet matrices that own their data.

pub fn new(shape: (usize, usize)) -> Self

Create a new triplet matrix of shape (nb_rows, nb_cols)

pub fn with_capacity(shape: (usize, usize), cap: usize) -> Self

Create a new triplet matrix of shape (nb_rows, nb_cols), and pre-allocate cap elements on the backing storage

pub fn from_triplets(
shape: (usize, usize),
row_inds: Vec<I>,
col_inds: Vec<I>,
data: Vec<N>
) -> Self

Create a triplet matrix from its raw components. All arrays should have the same length.

Panics

if the arrays don’t have the same length

if either the row or column indices are out of bounds.

pub fn add_triplet(&mut self, row: usize, col: usize, val: N)

Append a non-zero triplet to this matrix.

pub fn reserve(&mut self, cap: usize)

Reserve cap additional non-zeros

pub fn reserve_exact(&mut self, cap: usize)

Reserve exactly cap non-zeros

impl<N, I: SpIndex, IStorage, DStorage> TriMatBase<IStorage, DStorage> where
IStorage: Deref<Target = [I]>,
DStorage: Deref<Target = [N]>,

Common methods shared by all variants of triplet matrices

pub fn rows(&self) -> usize

The number of rows of the matrix

pub fn cols(&self) -> usize

The number of cols of the matrix

pub fn shape(&self) -> (usize, usize)

The shape of the matrix, as a (rows, cols) tuple

pub fn nnz(&self) -> usize

The number of non-zero entries

pub fn row_inds(&self) -> &[I]

The non-zero row indices

pub fn col_inds(&self) -> &[I]

The non-zero column indices

pub fn data(&self) -> &[N]

The non-zero values

pub fn find_locations(&self, row: usize, col: usize) -> Vec<TripletIndex>

Find all non-zero entries at the location given by row and col

pub fn transpose_view(&self) -> TriMatViewI<'_, N, I>

Get a transposed view of this matrix

pub fn triplet_iter(&self) -> TriMatIter<Iter<'_, I>, Iter<'_, I>, Iter<'_, N>>

Get an iterator over non-zero elements stored by this matrix

pub fn to_csc<Iptr: SpIndex>(&self) -> CsMatI<N, I, Iptr> where
N: Clone + Num,

Create a CSC matrix from this triplet matrix

pub fn to_csr<Iptr: SpIndex>(&self) -> CsMatI<N, I, Iptr> where
N: Clone + Num,

Create a CSR matrix from this triplet matrix

pub fn view(&self) -> TriMatViewI<'_, N, I>

impl<'a, N, I: SpIndex> TriMatBase<&'a [I], &'a [N]>

pub fn triplet_iter_rbr(
&self
) -> TriMatIter<Iter<'a, I>, Iter<'a, I>, Iter<'a, N>>

Get an iterator over non-zero elements stored by this matrix

Reborrowing version of triplet_iter().

impl<N, I: SpIndex, IStorage, DStorage> TriMatBase<IStorage, DStorage> where
IStorage: DerefMut<Target = [I]>,
DStorage: DerefMut<Target = [N]>,

pub fn set_triplet(
&mut self,
TripletIndex: TripletIndex,
row: usize,
col: usize,
val: N
)

Replace a non-zero value at the given index. Indices can be obtained using find_locations.

pub fn view_mut(&mut self) -> TriMatViewMutI<'_, N, I>