Trait sprs::DenseVector

pub trait DenseVector: Sealed {
    type Owned;
    type Scalar;
    fn dim(&self) -> usize;

    fn index(&self, idx: usize) -> &Self::Scalar;

    fn zeros(dim: usize) -> Self::Owned;

    fn to_owned(&self) -> Self::Owned;
}

A trait for types representing dense vectors, useful for expressing algorithms such as sparse-dense dot product, or linear solves.

This trait is sealed, and cannot be implemented outside of the sprs crate.

Associated Types

type Owned

type Scalar

Required methods

fn dim(&self) -> usize

The dimension of the vector

fn index(&self, idx: usize) -> &Self::Scalar

Random access to an element in the vector.

Panics

If the index is out of bounds

fn zeros(dim: usize) -> Self::Owned

Create an owned version of this dense vector type, filled with zeros

fn to_owned(&self) -> Self::Owned

Copies this vector into an owned version