Trait sprs::DenseVectorMut

pub trait DenseVectorMut: DenseVector {
    fn index_mut(&mut self, idx: usize) -> &mut Self::Scalar;
}

Trait for dense vectors that can be modified, useful for expressing algorithms which compute a resulting dense vector, such as solvers.

This trait is sealed, and cannot be implemented outside of the sprs crate.

Required methods

fn index_mut(&mut self, idx: usize) -> &mut Self::Scalar

Random mutable access to an element in the vector.

Panics

If the index is out of bounds