Type Definition sprs::CsVecI

type CsVecI<N, I = usize> = CsVecBase<Vec<I>, Vec<N>, N, I>;

Implementations

impl<N, I: SpIndex> CsVecI<N, I>

Methods operating on owning sparse vectors

pub fn empty(dim: usize) -> Self

Create an empty CsVec, which can be used for incremental construction

pub fn append(&mut self, ind: usize, val: N)

Append an element to the sparse vector. Used for incremental building of the CsVec. The append should preserve the structure of the vector, ie the newly added index should be strictly greater than the last element of indices.

Panics

Panics if ind is lower or equal to the last element of self.indices()

Panics if ind is greater than self.dim()

pub fn reserve(&mut self, size: usize)

Reserve size additional non-zero values.

pub fn reserve_exact(&mut self, exact_size: usize)

Reserve exactly exact_size non-zero values.

pub fn clear(&mut self)

Clear the underlying storage