Module sprs::vec

Structs

CsVecBase
A sparse vector, storing the indices of its non-zero data.

NnzIndex
Hold the index of a non-zero element in the compressed storage

NnzOrZip
An iterator over the non zeros of either of two vector iterators, ordered, such that the sum of the vectors may be computed

VectorIterator
An iterator over the non-zero elements of a sparse vector

VectorIteratorMut
An iterator over the mutable non-zero elements of a sparse vector

Enums

NnzEither

Traits

IntoSparseVecIter
Trait for types that can be iterated as sparse vectors

SparseIterTools

VecDim
A trait to represent types which can be interpreted as vectors of a given dimension.

Type Definitions

CsVec

CsVecView

CsVecViewMut