Trait sprs::vec::IntoSparseVecIter

pub trait IntoSparseVecIter<'a, N: 'a> {
    type IterType;
    fn into_sparse_vec_iter(
        self
    ) -> <Self as IntoSparseVecIter<'a, N>>::IterType
    where
        <Self as IntoSparseVecIter<'a, N>>::IterType: Iterator<Item = (usize, &'a N)>;

    fn dim(&self) -> usize;

    fn is_dense(&self) -> bool { ... }

    fn index(self, idx: usize) -> &'a N
    where
        Self: Sized,
    { ... }
}

Trait for types that can be iterated as sparse vectors

Associated Types

type IterType

Required methods

fn into_sparse_vec_iter(self) -> <Self as IntoSparseVecIter<'a, N>>::IterType where
<Self as IntoSparseVecIter<'a, N>>::IterType: Iterator<Item = (usize, &'a N)>,

Transform self into an iterator that yields (usize, &N) tuples where the usize is the index of the value in the sparse vector. The indices should be sorted.

fn dim(&self) -> usize

The dimension of the vector

Provided methods

fn is_dense(&self) -> bool

Indicator to check whether the vector is actually dense

fn index(self, idx: usize) -> &'a N where
Self: Sized,

Random access to an element in the vector.

Panics

if the vector is not dense

if the index is out of bounds