Struct sprs::stack::DStack

pub struct DStack<I> {
    // some fields omitted
}

A double stack of fixed capacity, growing from the left to the right or conversely.

Used in sparse triangular / sparse vector solves, where it is guaranteed that the two parts of the stack cannot overlap.

Methods

impl<I> DStack<I> where I: Copy

fn with_capacity(n: usize) -> DStack<I> where I: Default

Create a new double stacked suited for containing at most n elements

fn capacity(&self) -> usize

Capacity of the dstack

fn is_left_empty(&self) -> bool

Test whether the left stack is empty

fn is_right_empty(&self) -> bool

Test whether the right stack is empty

fn push_left(&mut self, value: I)

Push a value on the left stack

fn push_right(&mut self, value: I)

Push a value on the right stack

fn pop_left(&mut self) -> Option<I>

Pop a value from the left stack

fn pop_right(&mut self) -> Option<I>

Pop a value from the right stack

fn len_right(&self) -> usize

Number of right elements this double stack contains

fn clear_right(&mut self)

Clear the right stack

fn clear_left(&mut self)

Clear the left stack

fn iter_right<'a>(&'a self) -> Iter<'a, I>

Iterates along the right stack without removing items

fn push_left_on_right(&mut self)

Push the values of the left stack onto the right stack.

fn push_right_on_left(&mut self)

Push the values of the right stack onto the left stack.